+++ extensions/mole/src/installer.tsx
@@ -5,6 +5,7 @@ import { join } from "path";
 import { useState, useEffect, useMemo } from "react";
 import { getMolePathSafe } from "./utils/mole";
 import { formatBytes } from "./utils/parsers";
+import { MoleNotInstalled } from "./components/MoleNotInstalled";
 
 interface InstallerFile {
   name: string;
@@ -71,15 +72,7 @@ export default function CleanInstallers() {
   const molePath = useMemo(() => getMolePathSafe(), []);
 
   if (!molePath) {
-    return (
-      <List>
-        <List.EmptyView
-          title="Mole Not Installed"
-          description="Install Mole to use this extension: brew install mole"
-          icon={Icon.ExclamationMark}
-        />
-      </List>
-    );
+    return <MoleNotInstalled />;
   }
 
   return <InstallerView />;