+++ extensions/mole/src/optimize.tsx
@@ -3,20 +3,13 @@ import { useCachedPromise, showFailureToast } from "@raycast/utils";
 import { useMemo } from "react";
 import { getMolePathSafe, runMole } from "./utils/mole";
 import { parseOptimizeDryRun } from "./utils/parsers";
+import { MoleNotInstalled } from "./components/MoleNotInstalled";
 
 export default function OptimizeSystem() {
   const molePath = useMemo(() => getMolePathSafe(), []);
 
   if (!molePath) {
-    return (
-      <List>
-        <List.EmptyView
-          title="Mole Not Installed"
-          description="Install Mole to use this extension: brew install mole"
-          icon={Icon.ExclamationMark}
-        />
-      </List>
-    );
+    return <MoleNotInstalled />;
   }
 
   return <OptimizeView />;