+++ extensions/mole/src/uninstall.tsx
@@ -5,6 +5,7 @@ import { showFailureToast } from "@raycast/utils";
 import { useState, useEffect, useMemo } from "react";
 import { getMolePathSafe } from "./utils/mole";
 import { formatBytes } from "./utils/parsers";
+import { MoleNotInstalled } from "./components/MoleNotInstalled";
 import { execFile } from "child_process";
 
 interface AppInfo {
@@ -414,15 +415,7 @@ export default function UninstallApp() {
   const molePath = useMemo(() => getMolePathSafe(), []);
 
   if (!molePath) {
-    return (
-      <List>
-        <List.EmptyView
-          title="Mole Not Installed"
-          description="Install Mole to use this extension: brew install mole"
-          icon={Icon.ExclamationMark}
-        />
-      </List>
-    );
+    return <MoleNotInstalled />;
   }
 
   return <UninstallView />;